Improved Error Checking
Got FMOD Error Checking included and sorted out, good info! Thought about making it only in Debug, but the console output is pretty much debug info for the user ANYWAY. Maybe I'll do a pass once it's buttoned up and hold back some of the other debug spew.

Also told it to play the Title Theme event once the bot starts. This will probably come out of the speakers, but we've also encountered a bug where the banks aren't being properly copied to the build location.