Partial playback fix
Playback is now at the correct pitch. The speed, however, is all messed up. When playing sine wave, the buffer fills slower than the playback and it starves after half a second or so of playback. When playing the music, for some reason, it whips along at top speed as if one were scrubbing quickly through the file.