Workaround for Asset Capitalization ref bug
Funky thing where some of the assets referenced in MUS_SpringBlues were a version of an asset with a capital letter (ex: Harp) but the Asset Bin only saw an uncapitalized version (ex: harp). These were both the same file but FMOD Studio didn't know this, and so the Asset Bin thought the uncapitalized version was a separate unimported asset.