Fix for warnings with signed Ints in for loops
In places where we must use a for loop and iterate up to the Size of a vector, we were getting warnings about a signed / unsigned mismatch, due to std::vector::size() returning as an unsigned integer. These have now been addressed explicitly.

Also rebuilt the 32-bit version for the first time since renaming the project to Troubadour-Tabletop-Audio, which required some automatic changes to the .recipe file and others.