Fix for crash when loading banks with complex FMOD paths
A fix for an insidious bug that exposed flaws in my bank-loading and unloading process. When testing, we simply build banks straight to the soundbanks folder, but if the FMOD Studio path of the bank (bank://MyBank) has a folder in it (bank://MyFolder/MyBank), it causes problems.

Our fix: simply don't unload anything, and do less checks when loading, instead relying on FMOD Studio to tell us what's up using FMOD_RESULT's. This also avoids issues around somebody renaming banks on-the-fly, although it remains to be seen what happens if FMOD tries to stream a sound from a renamed bank.

Overall, we should push people to quit and restart the bot if they make bank changes.